API-373: fix a bug when parent is not set
When you import a product with no parent columns and you try to update
the common attribute, it updates the common attribute on the last level
(aka on the product variant level) but it should not. So you have
invalid data in the product values. It does not appears on the UI
because we fetch the data directly from the product model.
To avoid it we force set the parent, like this the attribute filter
knows it's a variant and skip the common attributes.